+++ src/components/index.js
@@ -84,5 +84,10 @@ export Types from './manage/Types/Types';
 export WysiwygWidget from './manage/Widgets/WysiwygWidget';
 export Workflow from './manage/Workflow/Workflow';
 
+export EditDescriptionTile from './manage/Tiles/Description/Edit';
 export EditTitleTile from './manage/Tiles/Title/Edit';
 export EditTextTile from './manage/Tiles/Text/Edit';
+
+export ViewDescriptionTile from './manage/Tiles/Description/View';
+export ViewTitleTile from './manage/Tiles/Title/View';
+export ViewTextTile from './manage/Tiles/Text/View';