+++ src/components/manage/Edit/Edit.jsx
@@ -48,6 +48,7 @@ const messages = defineMessages({
     content: state.content.data,
     schema: state.schema.schema,
     getRequest: state.content.get,
+    schemaRequest: state.schema,
     updateRequest: state.content.update,
     pathname: props.location.pathname,
     returnUrl: props.location.query.return_url,
@@ -78,6 +79,10 @@ export class EditComponent extends Component {
     getContent: PropTypes.func.isRequired,
     getSchema: PropTypes.func.isRequired,
     updateRequest: PropTypes.shape({
+      loading: PropTypes.bool,
+      loaded: PropTypes.bool,
+    }).isRequired,
+    schemaRequest: PropTypes.shape({
       loading: PropTypes.bool,
       loaded: PropTypes.bool,
     }).isRequired,
@@ -140,6 +145,13 @@ export class EditComponent extends Component {
     if (this.props.getRequest.loading && nextProps.getRequest.loaded) {
       this.props.getSchema(nextProps.content['@type']);
     }
+    if (this.props.schemaRequest.loading && nextProps.schemaRequest.loaded) {
+      if (nextProps.schema.properties.tiles) {
+        this.setState({
+          visual: true,
+        });
+      }
+    }
     if (this.props.updateRequest.loading && nextProps.updateRequest.loaded) {
       browserHistory.push(
         this.props.returnUrl || getBaseUrl(this.props.pathname),
@@ -185,7 +197,7 @@ export class EditComponent extends Component {
    * @returns {string} Markup for the component.
    */
   render() {
-    if (this.props.schema && this.props.content) {
+    if (this.props.schemaRequest.loaded && this.props.content) {
       return (
         <div id="page-edit">
           <Helmet
@@ -238,17 +250,19 @@ export class EditComponent extends Component {
                       title={this.props.intl.formatMessage(messages.save)}
                     />
                   </a>
-                  <a className="item" onClick={() => this.onToggleVisual()}>
-                    <Icon
-                      name={this.state.visual ? 'tasks' : 'block layout'}
-                      size="big"
-                      title={this.props.intl.formatMessage(
-                        this.state.visual
-                          ? messages.properties
-                          : messages.visual,
-                      )}
-                    />
-                  </a>
+                  {this.props.schema.properties.tiles && (
+                    <a className="item" onClick={() => this.onToggleVisual()}>
+                      <Icon
+                        name={this.state.visual ? 'tasks' : 'block layout'}
+                        size="big"
+                        title={this.props.intl.formatMessage(
+                          this.state.visual
+                            ? messages.properties
+                            : messages.visual,
+                        )}
+                      />
+                    </a>
+                  )}
                   <a className="item" onClick={() => this.onCancel()}>
                     <Icon
                       name="close"