+++ src/components/theme/View/TabularView.jsx
@@ -23,7 +23,7 @@ const TabularView = ({ content }) => (
       <header>
         <h1 className="documentFirstHeading">{content.title}</h1>
         {content.description && (
-          <p className="description">{content.description}</p>
+          <p className="documentDescription">{content.description}</p>
         )}
       </header>
       <section id="content-core">